+++ R/is_baysian_grid.R
@@ -20,7 +20,7 @@
 
 #' @keywords internal
 .is_baysian_grid.slopes <- function(x) {
-  insight::check_if_installed("marginaleffects")
+  insight::check_if_installed("marginaleffects", minimum_version = "0.29.0")
   !is.null(suppressWarnings(marginaleffects::get_draws(x, "PxD")))
 }