+++ R/utils.R
@@ -219,7 +219,7 @@
 #' @keywords internal
 .get_marginaleffects_draws <- function(object) {
   # errors and checks are handled by marginaleffects
-  insight::check_if_installed("marginaleffects", minimum_version = "0.24.0")
+  insight::check_if_installed("marginaleffects", minimum_version = "0.29.0")
   as.data.frame(marginaleffects::get_draws(object, shape = "DxP"))
 }